Move loadPartialManifest to separate file
